renamed app to "djsms".
this means that the project can't be named "djsms" (thanks django), so
you might need to rename your clone directory. i'm not worried by this
right now, since it's just a toy.

i also took the opportunity to remove the relative imports, since they
cause me so much trouble with nose and fudge. it's hard to stub things
out when you can't predict the name by which they will be imported.